+++ src/app/numbergame/numbergame.component.ts
@@ -194,7 +194,7 @@ export class NumbergameComponent implements OnInit {
     this.scatterChartData.forEach(x => {
       const data: any[] = x.data as any[];
       this.numberGame
-        .map(z => ({ x: parseInt(z.age), y: parseInt(z.score) })).forEach(a => data.push(a))
+        .map(z => ({ x: parseInt(z.age), y: parseFloat(z.score) })).forEach(a => data.push(a))
     })
   }
 }
